More methods

+++ src/main/java/com/revature/koality/service/PlaylistServiceImpl.java
@@ -8,15 +8,12 @@ import org.springframework.stereotype.Service;
 
 import com.revature.koality.bean.Playlist;
 import com.revature.koality.bean.Track;
-import com.revature.koality.dao.CustomerDAO;
-import com.revature.koality.dao.CustomerDAOImpl;
 import com.revature.koality.dao.PlaylistDAO;
 import com.revature.koality.dao.PlaylistDAOImpl;
 
 @Service("playlistServiceImpl")
 public class PlaylistServiceImpl implements PlaylistService {
 
-	private CustomerDAO cd;
 	private PlaylistDAO pld;
 
 	public PlaylistServiceImpl() {
@@ -24,30 +21,9 @@ public class PlaylistServiceImpl implements PlaylistService {
 
 	}
 
-	public PlaylistServiceImpl(CustomerDAOImpl cd) {
-		super();
-		this.cd = cd;
-	}
-
 	public PlaylistServiceImpl(PlaylistDAOImpl pld) {
 		super();
 		this.pld = pld;
-	}
-
-	public PlaylistServiceImpl(PlaylistDAOImpl pld, CustomerDAOImpl cd) {
-		super();
-		this.cd = cd;
-		this.pld = pld;
-	}
-
-	public CustomerDAO getCd() {
-		return cd;
-	}
-
-	@Autowired
-	@Qualifier("customerDAOImpl")
-	public void setCd(CustomerDAO cd) {
-		this.cd = cd;
 	}
 
 	public PlaylistDAO getPld() {
@@ -61,38 +37,74 @@ public class PlaylistServiceImpl implements PlaylistService {
 	}
 
 	@Override
-	public List<Track> getPurchasedTracks(int customerId) {
-		return cd.getAllTracksByCustomerId(customerId);
-	}
-
-	@Override
 	public int createPlaylist(String playlistName, int customerId) {
 		return pld.addPlaylist(playlistName, customerId);
 	}
 
 	@Override
-	public boolean deletePlaylist(int playlistId) {
-		return pld.deletePlaylist(playlistId);
+	public boolean deletePlaylist(int playlistId, int customerId) {
+
+		if (pld.isOwner(playlistId, customerId)) {
+			return pld.deletePlaylist(playlistId);
+		} else {
+			return false;
+		}
+
 	}
 
 	@Override
-	public int addTracksToPlaylist(int playlistId, List<Integer> trackIdList) {
-		return pld.addTracksToPlaylist(playlistId, trackIdList);
+	public int addTracksToPlaylist(int playlistId, List<Integer> trackIdList, int customerId) {
+
+		if (pld.isOwner(playlistId, customerId)) {
+			return pld.addTracksToPlaylist(playlistId, trackIdList);
+		} else {
+			return -1;
+		}
+
 	}
 
 	@Override
-	public int removeTracksFromPlaylist(int playlistId, List<Integer> trackIdList) {
-		return pld.removeTracksFromPlaylist(playlistId, trackIdList);
+	public int removeTracksFromPlaylist(int playlistId, List<Integer> trackIdList, int customerId) {
+
+		if (pld.isOwner(playlistId, customerId)) {
+			return pld.removeTracksFromPlaylist(playlistId, trackIdList);
+		} else {
+			return -1;
+		}
+
 	}
 
 	@Override
 	public List<Playlist> getAllPlaylists(int customerId) {
-		return pld.getAllPlaylistsByCustomerId(customerId);
+
+		List<Playlist> playlistList = pld.getAllPlaylistsByCustomerId(customerId);
+		playlistList.forEach(pl -> pl.truncate(true));
+		return playlistList;
+
 	}
 
 	@Override
-	public Playlist selectPlaylist(int playlistId) {
-		return pld.getPlaylistById(playlistId);
+	public Playlist selectPlaylist(int playlistId, int customerId) {
+
+		if (pld.isOwner(playlistId, customerId)) {
+			return pld.getPlaylistById(playlistId);
+		} else {
+			return null;
+		}
+
+	}
+
+	@Override
+	public List<Track> getPlaylistTracks(int playlistId, int customerId) {
+
+		if (pld.isOwner(playlistId, customerId)) {
+			List<Track> trackList = pld.getTracksFromPlaylist(playlistId);
+			trackList.forEach(t -> t.truncate(false));
+			return trackList;
+		} else {
+			return null;
+		}
+
 	}
 
 }
